Added get members of a classroom

+++ src/main/java/com/example/hitproduct/repository/ClassroomRepository.java
@@ -12,6 +12,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClassroomRepository extends JpaRepository<Classroom, Long> {
+public interface ClassroomRepository extends JpaRepository<Classroom, Integer> {
     boolean existsByName(String name);
 }
